fix:添加mapper

+++ marketing-cs/src/main/java/com/br/marketing/mapper/CarClueInfoMapperBase.java
@@ -2,12 +2,11 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.CarClueInfo;
 import com.br.marketing.entity.CarClueInfoExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-import java.util.List;
-
 public interface CarClueInfoMapperBase {
-    int countByExample(CarClueInfoExample example);
+    long countByExample(CarClueInfoExample example);
 
     int deleteByExample(CarClueInfoExample example);
